biocmask Data Mask Object

An R6 Object that tracks bindings of a list-like object. This includes DFrame objects. There are several inherited environments that the data is stored within.

Environments:

.shared_env –> curr_group_ctx –> foreign –> lazy –> chops –> active_mask

.shared_env : environment provided at initialization. This may be shared with multiple other BiocDataMasks.

curr_group : Currently not used.

foreign : space to put foreign bindings, i.e. object unrelated to .data provided at initialization. This space is currently used to place the pronouns into related contexts.

lazy : A strict lazy binding to the data within .data. This binding is made only at initialization.

chops : lazy data but chopped into list by .indices. New bindings for this BiocDataMask context are expected to be in a "chopped" format and are assigned here.

active_mask : An active binding to chops in which the proper list index is used depending on the current group context. The current group context is at this moment determined by the .shared_env NOT the curr_group. I have plans to remove the curr_group environment.

Active bindings

environments

the hierarchy of environments for this mask

names

the associated names of data in mask

ptype

get zero-length slices of underlying data

added

newly added names to the mask holds grouping information for this object foreign bindings

Methods

Public methods

biocmask$new()

biocmask$on_bind()

biocmask$bind()

biocmask$unchop()

biocmask$results()

biocmask$eval()

biocmask$clone()

Method new()

Create a biocmask from .data. .data is chopped by .indices, and environments are built from .env

Usage

biocmask$new(.data, .indices = NULL, .env)

Arguments

.data

a named list like object to create a mask

.indices

the indices that will be used to chop .data

.env

an environment that the resulting mask will be built from.

Method on_bind()

appends a callback function that is executed after a value is bound to this mask. Mainly used to inform other masks of new values

Usage

biocmask$on_bind(.fun)

Arguments

.fun

a function created from add_bind()

Method bind()

binds value to an name within the chops environment.

Usage

biocmask$bind(name, value)

Arguments

name

a character scalar

value

results from $eval in the form of chops

Method unchop()

Usage

biocmask$unchop(name)

Arguments

name

name of binding to retrieve and unchop

Method results()

Usage

biocmask$results()

Returns

named list of evaluated expression, unchopped

Method eval()

evaluates a quoted expression within a new datamask

Usage

biocmask$eval(quo, env = caller_env())

Arguments

quo

a quosure to evaluate

env

an environment to search after mask

Method clone()

The objects of this class are cloneable with this method.

Usage

biocmask$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.