biocmask for SummarizedExperiment assays()

A more specialized version of the biocmask R6 object for the assays list object. This includes chopping and unchopping of matrix like objects.

Super class

biocmask::biocmask -> biocmask_assay

Methods

Public methods

biocmask_assay$new()

biocmask_assay$unchop()

biocmask_assay$clone()

Inherited methods

biocmask::biocmask$bind()

biocmask::biocmask$eval()

biocmask::biocmask$on_bind()

biocmask::biocmask$results()

Method new()

Create a biocmask from .data. .data is chopped by .indices, and environments are built from .env

Usage

biocmask_assay$new(.data, .indices, .env, .nrow, .ncol)

Arguments

.data

a named list like object to create a mask

.indices

the indices that will be used to chop .data

.env

an environment that the resulting mask will be built from.

.nrow, .ncol

the number of rows and columns of each element of .data respectively

Method unchop()

unchop data within the mask, returns a matrix

Usage

biocmask_assay$unchop(name)

Arguments

name

name of binding to retrieve and unchop

Method clone()

The objects of this class are cloneable with this method.

Usage

biocmask_assay$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.