biocmask Data Mask Manager

This object organizes serveral biocmasks, allowing expressions to be evaluated in different contexts.

Active bindings

ctx

get and set the current context

ctx_mask

get the current context biocmask

n_groups

get the current context biocmask group size

group_id

get and set the current context biocmask group id

masks

get the private list of masks

extended

other environments extended from a context mask.

Methods

Public methods

biocmask_manager$new()

biocmask_manager$along_ctx()

biocmask_manager$eval()

biocmask_manager$results()

biocmask_manager$clone()

Method new()

Usage

biocmask_manager$new(.data, .masks, .ctx_env, .extended_env)

Arguments

.data

Original data

.masks

list of named biocmask objects

.ctx_env

shared context environment

.extended_env

other extended environments

Method along_ctx()

provides a sequence for iterate over the groups

Usage

biocmask_manager$along_ctx()

Method eval()

eval an expression in the current context

Usage

biocmask_manager$eval(quo, name, env = caller_env())

Arguments

quo

a quosure or quoted expression

name

the resulting name to bind

env

an environment

Returns

returns evaluated quo in the form of a chop

Method results()

collects the evaluated results with biocmasks

Usage

biocmask_manager$results()

Returns

named list for each mask containing named list of evaluated expressions.

Method clone()

The objects of this class are cloneable with this method.

Usage

biocmask_manager$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.