biocmask: Create S4 BioConductor data masks for dplyr functions

The package provides `rlang` data masks for SummarizedExperiment class. The enables the evaluation of unquoted expression in different contexts of the SummarizedExperiment object with optional access to other contexts. The goal for `biocmask` is for evaluation to feel like a data.frame object without ever needing to unwind to a rectangular data.frame.

Value

API for using S4 classes with rlang data masks

See also

Author

Maintainer: Justin Landis jtlandis314@gmail.com (ORCID)

Authors:

Michael Love michaelisaiahlove@gmail.com (ORCID)