contextual biocmask pronouns

biocmask utilizes its own version of rlang::.data pronouns. These may be used to gain access to other evaluation contexts for a managed set of data-masks.

Similar to rlang::.data, biocmask::.assays and other exported pronouns are exported to pass R CMD Checks. When using a biocmask within your package, import the associated pronoun from biocmask but only use the fully unqualified name, .assays, .assays_asis, etc.

Usage

.assays

.assays_asis

.rows

.rows_asis

.cols

.cols_asis

Format

An object of class NULL of length 0.

An object of class NULL of length 0.

An object of class NULL of length 0.

An object of class NULL of length 0.

An object of class NULL of length 0.

An object of class NULL of length 0.