Mutate a SummarizedExperiment object

Mutate a SummarizedExperiment object under an data mask. Unlike a few other dplyr implementations, all contextual evaluations of mutate() for SummarizedExperiment are valid.

Usage

# S3 method for class 'SummarizedExperiment'
mutate(.data, ...)

Arguments

.data

a SummarizedExperiment object

...

expressions to evaluate

Value

SummarizedExperiment object

Examples

mutate(se_simple,
    counts_1 = counts + 1,
    logp_counts = log(counts_1),
    # access assays context with ".assays" pronoun,
    # note that assays are sliced into a list to 
    # fit dimensions of cols context
    cols(sum = purrr::map_dbl(.assays$counts, sum)),
    # access assays context "asis" with the same pronoun
    # but with a "_asis" suffix.
    rows(sum = rowSums(.assays_asis$counts))
)
#> Error in dplyr::mutate(as_tibble(.data, skip_GRanges = skip_GRanges),     ...): ℹ In argument: `cols(sum = purrr::map_dbl(.assays$counts, sum))`.
#> Caused by error in `cols()`:
#> ! `cols()` is a sentinal function for SummarizedExperiment dplyr verbs