Filter SummarizedExperiment

The filter() function is sused to subset an object, returing the observations that satisfy your conditions. An observation must return TRUE for all conditions within a context to be retained. Note, to guarantee a valid SummarizedExperiment is returned, filtering in the assays evaluation context is disabled.

Usage

# S3 method for class 'SummarizedExperiment'
filter(.data, ..., .preserve = FALSE)

Arguments

.data

A SummarizedExperiment object

...

conditions to filter on. These must be wrapped in cols() and or rows()

.preserve

Relevant when the .data input is grouped. If .preserve = FALSE (the default), the grouping structure is recalculated based on the resulting data, i.e. the number of groups may change.

Value

SummarizedExperiment Object

Examples

# example code
filter(se_simple,
       rows(length > 30),
       cols(condition == "drug"))
#> Error in dplyr::filter(as_tibble(.data, skip_GRanges = TRUE), ..., .preserve = .preserve): ℹ In argument: `rows(length > 30)`.
#> Caused by error in `rows()`:
#> ! `rows()` is a sentinal function for SummarizedExperiment dplyr verbs
 
filter(se_simple,
       rows(rowSums(.assays_asis$counts) > 40),
       cols(colSums(.assays_asis$counts) < 50))
#> Error in dplyr::filter(as_tibble(.data, skip_GRanges = TRUE), ..., .preserve = .preserve): ℹ In argument: `rows(rowSums(.assays_asis$counts) > 40)`.
#> Caused by error in `rows()`:
#> ! `rows()` is a sentinal function for SummarizedExperiment dplyr verbs

# assay context is disabled
filter(se_simple,
       counts > 12) |> try()
#> tidySummarizedExperiment says: The resulting data frame is not rectangular (all genes for all samples), a tibble is returned for independent data analysis.
#> # A tibble: 8 × 9
#>   .feature .sample counts logcounts sample condition gene  length direction
#>   <chr>    <chr>    <int>     <dbl> <chr>  <chr>     <chr>  <int> <chr>    
#> 1 row_1    col_1       14      2.64 s1     cntrl     g1         1 -        
#> 2 row_2    col_1       19      2.94 s1     cntrl     g2        24 +        
#> 3 row_3    col_1       16      2.77 s1     cntrl     g3        60 +        
#> 4 row_5    col_1       18      2.89 s1     cntrl     g5        37 +        
#> 5 row_4    col_2       17      2.83 s2     cntrl     g4        39 -        
#> 6 row_5    col_2       13      2.56 s2     cntrl     g5        37 +        
#> 7 row_3    col_3       15      2.71 s3     drug      g3        60 +        
#> 8 row_3    col_4       20      3.00 s4     drug      g3        60 +        

# convert to `data.frame` first
as.data.frame(se_simple) |>
  filter(counts > 12)
#>   .features .samples counts logcounts gene length direction sample condition
#> 1     row_1    col_1     14  2.639057   g1      1         -     s1     cntrl
#> 2     row_2    col_1     19  2.944439   g2     24         +     s1     cntrl
#> 3     row_3    col_1     16  2.772589   g3     60         +     s1     cntrl
#> 4     row_5    col_1     18  2.890372   g5     37         +     s1     cntrl
#> 5     row_4    col_2     17  2.833213   g4     39         -     s2     cntrl
#> 6     row_5    col_2     13  2.564949   g5     37         +     s2     cntrl
#> 7     row_3    col_3     15  2.708050   g3     60         +     s3      drug
#> 8     row_3    col_4     20  2.995732   g3     60         +     s4      drug