select assays, rowData, and colData names

Select one or more values from each context. By default omitting an expression for a context is the same as selecting NOTHING from that context.

The <tidy-select> implementation within biocmask is almost similar to dplyr except when used within the across() function. When used from accross(), the data provided to eval_select is a zero length slice of the data. This was an intentional choice to prevent the evaluation of potentionally expensive chopping operations for S4Vectors. This means that predicate function from where() will NOT be able to query the original data.

Usage

# S3 method for class 'SummarizedExperiment'
select(.data, ...)

Arguments

.data

a SummarizedExperiment object

...

<tidy-select> one or more selection expressions. Supports wrapping expressions within the <biocmask-contexts>.

Examples

# only keep assays, other contexts are dropped
select(se_simple, everything())
#> # A SummarizedExperiment-tibble abstraction: 20 × 9
#> # Features=5 | Samples=4 | Assays=counts, logcounts
#>    .feature .sample counts logcounts sample condition gene  length direction
#>    <chr>    <chr>    <int>     <dbl> <chr>  <chr>     <chr>  <int> <chr>    
#>  1 row_1    col_1       14     2.64  s1     cntrl     g1         1 -        
#>  2 row_2    col_1       19     2.94  s1     cntrl     g2        24 +        
#>  3 row_3    col_1       16     2.77  s1     cntrl     g3        60 +        
#>  4 row_4    col_1       11     2.40  s1     cntrl     g4        39 -        
#>  5 row_5    col_1       18     2.89  s1     cntrl     g5        37 +        
#>  6 row_1    col_2        8     2.08  s2     cntrl     g1         1 -        
#>  7 row_2    col_2        2     0.693 s2     cntrl     g2        24 +        
#>  8 row_3    col_2        6     1.79  s2     cntrl     g3        60 +        
#>  9 row_4    col_2       17     2.83  s2     cntrl     g4        39 -        
#> 10 row_5    col_2       13     2.56  s2     cntrl     g5        37 +        
#> 11 row_1    col_3        7     1.95  s3     drug      g1         1 -        
#> 12 row_2    col_3        1     0     s3     drug      g2        24 +        
#> 13 row_3    col_3       15     2.71  s3     drug      g3        60 +        
#> 14 row_4    col_3       10     2.30  s3     drug      g4        39 -        
#> 15 row_5    col_3       12     2.48  s3     drug      g5        37 +        
#> 16 row_1    col_4        9     2.20  s4     drug      g1         1 -        
#> 17 row_2    col_4        4     1.39  s4     drug      g2        24 +        
#> 18 row_3    col_4       20     3.00  s4     drug      g3        60 +        
#> 19 row_4    col_4        3     1.10  s4     drug      g4        39 -        
#> 20 row_5    col_4        5     1.61  s4     drug      g5        37 +        

# only keep rowData, other contexts are dropped
select(se_simple, rows(everything()))
#> Error in select_helper(as_tibble(.data %>% {    if (ncol(.) > 0)         .[1, 1, drop = FALSE]    else (.)}), ...): ℹ In argument: `rows(everything())`.
#> Caused by error in `rows()`:
#> ! `rows()` is a sentinal function for SummarizedExperiment dplyr verbs

select(se_simple, rows(where(is.numeric)))
#> Error in select_helper(as_tibble(.data %>% {    if (ncol(.) > 0)         .[1, 1, drop = FALSE]    else (.)}), ...): ℹ In argument: `rows(where(is.numeric))`.
#> Caused by error in `rows()`:
#> ! `rows()` is a sentinal function for SummarizedExperiment dplyr verbs

# Note on `where()` clause, all data is available within select
select(se_simple, rows(where(~any(grepl("-", .x)))))
#> Error in select_helper(as_tibble(.data %>% {    if (ncol(.) > 0)         .[1, 1, drop = FALSE]    else (.)}), ...): ℹ In argument: `rows(where(~any(grepl("-", .x))))`.
#> Caused by error in `rows()`:
#> ! `rows()` is a sentinal function for SummarizedExperiment dplyr verbs

# within an `across()`, only a zero-length slice avialble, so the
# `where()` predicate cannot access the data
mutate(se_simple,
       rows(
        across(where(~any(grepl("-", .x))),
               ~sprintf("%s foo", .x))))
#> Error in dplyr::mutate(as_tibble(.data, skip_GRanges = skip_GRanges),     ...): ℹ In argument: `rows(across(where(~any(grepl("-", .x))), ~sprintf("%s
#>   foo", .x)))`.
#> Caused by error in `rows()`:
#> ! `rows()` is a sentinal function for SummarizedExperiment dplyr verbs
# here is an acceptable usage of the `where()` predicate
mutate(se_simple,
       rows(
        across(where(is.character),
               ~sprintf("%s foo", .x))))
#> Error in dplyr::mutate(as_tibble(.data, skip_GRanges = skip_GRanges),     ...): ℹ In argument: `rows(across(where(is.character), ~sprintf("%s foo",
#>   .x)))`.
#> Caused by error in `rows()`:
#> ! `rows()` is a sentinal function for SummarizedExperiment dplyr verbs